feat: [ANDROAPP-5125] move SearchTeiModel from commons to app
