+++ app/src/main/java/org/dhis2/usescases/searchTrackEntity/SearchRepository.java
@@ -4,7 +4,6 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
 import org.dhis2.commons.data.EventViewModel;
-import org.dhis2.commons.data.SearchTeiModel;
 import org.dhis2.commons.data.tuples.Pair;
 import org.dhis2.commons.filters.FilterManager;
 import org.dhis2.commons.filters.sorting.SortingItem;
