+++ app/src/main/java/org/dhis2/usescases/searchTrackEntity/adapters/BaseTeiViewHolder.kt
@@ -9,10 +9,10 @@ import org.dhis2.bindings.paintAllEnrollmentIcons
 import org.dhis2.bindings.setAttributeList
 import org.dhis2.bindings.setStatusText
 import org.dhis2.commons.data.EnrollmentIconData
-import org.dhis2.commons.data.SearchTeiModel
 import org.dhis2.commons.date.toDateSpan
 import org.dhis2.commons.resources.ColorUtils
 import org.dhis2.databinding.ItemSearchTrackedEntityBinding
+import org.dhis2.usescases.searchTrackEntity.SearchTeiModel
 
 abstract class BaseTeiViewHolder(
     private val binding: ItemSearchTrackedEntityBinding,
