+++ app/src/main/java/org/dhis2/usescases/searchTrackEntity/adapters/SearchAdapterDiffCallback.kt
@@ -1,7 +1,7 @@
 package org.dhis2.usescases.searchTrackEntity.adapters
 
 import androidx.recyclerview.widget.DiffUtil
-import org.dhis2.commons.data.SearchTeiModel
+import org.dhis2.usescases.searchTrackEntity.SearchTeiModel
 
 class SearchAdapterDiffCallback() : DiffUtil.ItemCallback<SearchTeiModel>() {
     override fun areItemsTheSame(oldItem: SearchTeiModel, newItem: SearchTeiModel): Boolean {
