+++ app/src/main/java/org/dhis2/usescases/searchTrackEntity/adapters/SearchErrorViewHolder.kt
@@ -1,8 +1,8 @@
 package org.dhis2.usescases.searchTrackEntity.adapters
 
 import androidx.recyclerview.widget.RecyclerView
-import org.dhis2.commons.data.SearchTeiModel
 import org.dhis2.databinding.ItemSearchErrorBinding
+import org.dhis2.usescases.searchTrackEntity.SearchTeiModel
 
 class SearchErrorViewHolder(
     private val binding: ItemSearchErrorBinding,
